Make TileGrid linear of type u32

+++ src/asset/map.rs
@@ -55,7 +55,7 @@ impl<'a, R: 'a + Read> MapReader<'a, R> {
         self.reader.read_exact(&mut name[..]).unwrap();
 
         let entrance_pos_lin = self.reader.read_i32::<BigEndian>()?;
-        let entrance_pos = self.tile_grid.from_linear_inv(entrance_pos_lin);
+        let entrance_pos = self.tile_grid.from_linear_inv(entrance_pos_lin as u32);
         debug!("entrance_pos={} ({:?})", entrance_pos_lin, entrance_pos);
         let entrance_elevation = self.reader.read_u32::<BigEndian>()? as usize;
         assert!(entrance_elevation <= ELEVATION_COUNT);
@@ -358,7 +358,7 @@ impl<'a, R: 'a + Read> MapReader<'a, R> {
         let pos = if pos >= 0 {
             Some(ElevatedPoint {
                 elevation,
-                point: self.tile_grid.from_linear_inv(pos),
+                point: self.tile_grid.from_linear_inv(pos as u32),
             })
         } else {
             None
