+++ src/graphics/geometry/sqr.rs
@@ -45,7 +45,7 @@ impl TileGrid {
         self.pos = pos.into();
     }
 
-    pub fn set_pos_lin(&mut self, num: i32) {
+    pub fn set_pos_lin(&mut self, num: u32) {
         let lin = self.from_linear_inv(num);
         self.set_pos(lin);
     }
@@ -100,11 +100,11 @@ impl TileGrid {
     /// Rectangular to linear coordinates with `x` axis inverted.
     /// This method should be used when converting linears for use in the original assets
     /// (maps, scripts etc).
-    pub fn to_linear_inv(&self, p: impl Into<Point>) -> Option<i32> {
+    pub fn to_linear_inv(&self, p: impl Into<Point>) -> Option<u32> {
         let p = p.into();
         if self.is_in_bounds(p) {
             let x = self.width - 1 - p.x;
-            Some(self.width * p.y + x)
+            Some((self.width * p.y + x) as u32)
         } else {
             None
         }
@@ -113,9 +113,9 @@ impl TileGrid {
     /// Linear to rectangular coordinates with `x` axis inverted.
     /// This method should be used when converting linears for use in the original assets
     /// (maps, scripts etc).
-    pub fn from_linear_inv(&self, num: i32) -> Point {
-        let x = self.width - 1 - num % self.width;
-        let y = num / self.width;
+    pub fn from_linear_inv(&self, num: u32) -> Point {
+        let x = self.width - 1 - num as i32 % self.width;
+        let y = num as i32 / self.width;
         Point::new(x, y)
     }
